WestUtils Standard Update v0.1.1 This update mainly focuses on Graphics, which adds a new class hierarchy, the WG- prefixed classes, and a new drawing function. The WG- prefix is short for WestGraphics and mimics the "J- prefix" with the Swing objects, all this means is that these are classes that function with WestGraphics and define certain internal functions that make it easier to find what has been clicked and where the objects are in relation to other objects
Change log:
Added the WGDrawingObject class, which is essentially a class that defines defualt functionalities for all of these advanced graphic components. Currently just creates a getBounds function and the x, y, width, and height. It is a rectangle, but also not so it is not a direct subclass of the Rectangle2D.Double.
Added the WGLoadingBar class. This defines all of the necessary components to create a loading bar and has some defualt values if not supplied certain things.
Added the WGAnnouncementCard class and the corresponding fnction to WestGraphics. This object defines a title with a splitting line then a subtitle. The title and subtitle are required, but the line is not. As with the loading bar's object, this defines certain defaults and provides greater control over the object than just the draw version with parameters.