WestUtils Standard Update v0.1.2 Another great step towards the graphics initiative (By the end of the initiative animation, basic and advanced components, and internal management of resizing, clicking, and hovering should be completed). This step includes a font management device, that is mostly accurate, however further testing needs to be done to completely guarantee its accuracy. Additionally the button is back! And it is looking pretty good! Unfortunately the button is a simple class so it will have the most in-depth modularity to it, not that the other devices are not getting these capabilities, but it makes the most sense for the button to have certain key characteristics. Finally, added an internal resize listener class that allows for these object to resize themselves as they so desire.
Change log:
Added the WGButton class and its respective function that allows for a box with a text to appear on the screen. Not super extraordinary I know, but it has quite a few internal capabilities that make it more than it looks like.
Added the WGDrawingObjectResizeListener class. Now what is this very long named class you may ask? All this does is define certain key traits that all of the WGDrawingOjects need to be able to resize in the background. Without this class there wouldn't be a standardized way of making sure these objects resized properly
Added the WGFontHelper class. This is a static class that defines approximation functions to figure out about how big or small a font will be and of course the inverse where given a certain size (width or height) it will find out a proper sized font that would fit those bounds. This is especially useful in the WGButton class and other WGDrawingObject subclasses because it allows for the text to "fit" within the object and not stick out.
Modified all of the current classes to have a constructor that allows for the component to resize itself given certain percentages. This may be annoying, so there is always the definitions without that capabilities, just note that any other additions, like a click listener, will have to be manually entered. Although, that may be easier than the current 18 variable definition of an WGAnnouncementCard.
Fixed a bug where the loading bar's text was all over the place, versus being in the center of the component.
*Note: This version of graphicsUtilities no longer supports vertical loading bar text, therefore the text will always be horizontal until a new support gets developed.