WestUtils Standard Update v0.1.10
Made the logger hit v0.2! This creates a new system to handle unit tests. 

Miscellaneous changes to the graphics system, all stability and bug fixes. Added a few new definitions to certain classes that allow for a 2D rectangle to serve as the bounds. 

Added a new class, UnitTest, which serves to compare to strings, case sensitive, which allows for a expected result and a given result to be compared. This allows for complex functions to be compared against certain tests based upon the expected and returned results.