WestUtils Standard Update v0.2.8.3
Description:
This is the final commit to the v0.2 system as the system is changing
from a Swing-application base to a JavaFX-application base, which will
require massive and breaking changes to the UI. Unfortunately since
JavaFX is not backwards compatible with Swing (actually is not even
based on the same underlying system), we have to create a brand-new
system for this. Thankfully most of the items are one-to-one in
conversion, but there will still bee some problems during the conversion
phase. Once the conversion is completed there will be no going back, so
this is the last stable version of WestUtils for javax.Swing and
Java.awt systems. The other issue is that most of the WestUtil system is
STILL under maintenance, so there are STILL open issues with the system.
For instance WGDropDowns not drawing properly or the weird ways the new
mouse cursor system is interacting with shown and unshown UI components,
all of which should have been finished before the conversion. However,
this is on the record that the Graphics Utilities have always been
directly related to the development of Nightmare Game, and in order to
satisfy that requirement, the Graphics Utilities will only support the
system that Nightmare Game uses. (Although in the future there could be
a foresee-able backwards compatibility with Swing/AWT, but that is NOT
guaranteed!)

Changes:
 - Unfortunately I'm not sure of the changes because I have changed IDEs
and this one seems to be strange with the way it treats first commits
taken from GitHub, but I'm sure there were changes to the DataStructures
or small fixes to the WestGraphics system, but specifics are not
available