WestUtils Standard Update v0.3.0
Description:
Changed the entire system over to the JavaFX system utilizing the brand
new and improved way of handling the graphics. Since this is a directly
port expect errors! Also some of the original functionality of the
system has been changed! Some functions will not work! PLEASE DO NOT
CHANGE TO THIS VERSION WITHOUT FIRST CHANGING YOUR CODE TO WORK WITH
JAVAFX, OTHERWISE THIS WILL NOT WORK!! This is still an unstable
version, expect there to be other massive changes in the coming
updates!!

Changes:
 - Switched the entire system to use JavaFX
 - Added a helper class that models the swing FontMetrics. I did not
write this class!!