WestUtils Standard Update v0.3.3
Description:
Finished the fixes on the panes not scrolling issue. There should be no
more errors with scrolling, cross my fingers. Additionally, a long
standing enhancement to the system has been added, which are Ranges.
Currently ranges can be created with any generic E that extends the
Number class. This was done to ensure numeric values were being stored
in the range. Furthermore, the FileProcessor system has been updated
with a way to convert strings to Ranges

Changes:
 - Fixed #10
 - Fixed #17

Known Issues:
 - There are some known issues with the time complexity of the pane's
resizing and re-creating the scroll bounds. However, this problem is
minor in most cases, and is only annoying with certain resizings of the
screen (maximizing the screen as an example)