WestUtils Standard Update v0.3.4
Description:
Added WGDragDropBars and their respective clickListeners. This system is
mostly complete and mostly bug free. There are some residual issues
left; however, for the majority part they are not big issues. (Drag and
drop modes will be added later)

Changes:
 - Fixed #22

Known Issues:
 - There are some known issues with the time complexity of the pane's
resizing and re-creating the scroll bounds. However, this problem is
minor in most cases, and is only annoying with certain resizings of the
screen (maximizing the screen as an example)
 - The drag and drop bar can sometimes have a click "bleed through" to
underlying components. This is just an annoying issue, but not a system
breaking one