WestUtils Standard Update v0.3.4.1
Description:
Added WGOverlays, which are quick and easy ways to cover a UI component
and draw a color and text over it. Additionally fixed the issue with the
pane time efficiency issue, mostly. There is unfortunately still
something going on in the background that is causing some lag time;
however it is likely the problem is not caused by West Graphics, but is
just the nature of JavaFX.

Changes:
 - Added WGOverlays, which are quick and easy overlays with a textual
component
 - Fixed #19

Known Issues:
 - The drag and drop bar can sometimes have a click "bleed through" to
underlying components. This is just an annoying issue, but not a system
breaking one