WestUtils Standard Update v0.3.5
Description:
Finished changes to the drag and drop system to allow for Reorder Drag
Mode (RDM), which essentially just allows for one to reorder a list of
WGDrawingObjects. This requires a pane and that the objects attempting
to be swapped with are also drag drop able (Although they can be RDM or
some other mode). Additionally this issue adds a new variable to all of
the WGPanes!! This means that all definitions of WGPane will create red
errors now. (This is just a boolean value addition though).

Changes:
 - Fixed #25
 - WGPanes now require a new boolean value in all of the constructors
which defines whether the components within use relative positioning, or
absolute positioning (absolute as in relative to the parent Canvas, and
relative as in relative to the pane)

Known Issues:
 - The drag and drop bar can sometimes have a click "bleed through" to
underlying components. This is just an annoying issue, but not a system
breaking one
 - There is an issue with stacking functionality, so there is no way for
the system to determine one scroller from another. Therefore, stacked
scrollable panes and stacked drag and drops may not work properly until
bleed protection is created